feat(AttrEffect): 重构属性效果系统
- 移除 Ability 类的大部分代码，仅保留基本结构
- 新增 AbilitySet 类，用于管理能力集合
-重构 Attr 类，使用 double 类型替代 float
- 新增 AttrSet 类，用于管理属性集合
- 更新相关的场景和脚本，以适应新的属性效果系统

+++ Assets/Scripts/UI/UIPopupText.cs
@@ -39,7 +39,7 @@ namespace Kirara.UI
             _rectTransform = transform as RectTransform;
         }
 
-        public UIPopupText SetDamage(Transform follow, Vector3 localPos, float dmg, bool isCrit)
+        public UIPopupText SetDamage(Transform follow, Vector3 localPos, double dmg, bool isCrit)
         {
             _follow = follow;
             _localPos = localPos;
